+++ gesipan/src/main/java/com/zzangmin/gesipan/web/dto/post/PostUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.zzangmin.gesipan.web.dto;
+package com.zzangmin.gesipan.web.dto.post;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
